% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2HUpurifyEasyClusMetaData.R
\name{R2HUpurifyEasyClusMetaData}
\alias{R2HUpurifyEasyClusMetaData}
\title{Function to fix the meta data export tables from EasyClus, where certain column names shift due to our old instrument.}
\usage{
R2HUpurifyEasyClusMetaData(
  rawfilename = "./*.txt",
  separator = "XXX",
  writetocsv = T,
  newfilename = "./metadata_fixed.csv"
)
}
\arguments{
\item{working_directory, rawfile_folder, rawfile_identifier, writetofile, concatenatedfile_name, writetoexcel, excelfile_name, excelfile_sheet, apply_chlacalibration, setsinsupergroups, apply_lengthcalibration}{\code{(to be updated)}}
}
\value{
Returns the corrected table.
}
\description{
\code{-BETA-}   Ask Nico if you would like to use this function.
}
